Ensure die_function_type_is_method_type returns a class type die
When die_function_type_is_method returns true, it meanins that the
function die it's looking at is a member function.  In that case,
die_function_type_is_method must also return the class of which the
function is a member.  It appears that there are cases where this
die_function_type_is_method returns a *typedef* DIE (to a class DIE),
rather than a class DIE.

This results in the assertion violation below when loading the file
/usr/lib/libreoffice/program/libanalysislo.so from, e.g, the
libreoffice-calc-5.3.6.1-10.el7.i686 package:

    /home/dodji/git/libabigail/master/src/abg-dwarf-reader.cc:13846: abigail::ir::function_type_sptr abigail::dwarf_reader::build_function_type(abigail::dwarf_reader::read_context&, Dwarf_Die*, abigail::ir::class_or_union_sptr, size_t): Assertion `klass_type' failed.

This patch fixes that by peeling off the potential typedefs that might
be there.

	* src/abg-dwarf-reader.cc (die_peel_typedef): Define new static
	function.
	(die_function_type_is_method_type): Use the function above to peel
	the class die from potential typedefs wrapping it.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>