Fix race between runtestdefaultsupprs{py3.sh,.py}
runtestdefaultsupprspy3.sh just runs runtestdefaultsupprs.py using
python3.  So when both run in parallel, it can happen that they step
on their toes as they emit content into some files.

This patch commits makes it so that runtestdefaultsupprspy3.sh is
executed only if we are in the python3 mode, otherwise we just run
runtestdefaultsupprs.py, fixing the occasional failure that we see on
either runtestdefaultsupprs.py or runtestdefaultsupprspy3.sh.

	* tests/Makefile.am: Run runtestdefaultsupprspy3.sh if we are in
	python3 mode otherwise run runtestdefaultsupprs.py.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>