Allow use of python even when fedabipkgdiff is disabled
It turns out that when we disable fedabipkgdiff, we also disable
detection of python and the setting of the proper environment
variables that are expected by some makefiles down the road.  The
leads to the runtestdefaultsupprs.py test to be blocked at runtime
because the environement variable PYTHON is not defined.  This patch
fixes that.

	* configure.ac: Detect the presence and version of python even
	when fedabipkgdiff is disabled.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>