Add basic support for Fortran binaries
Apparently, the only DWARF TAG that we needed to support to allow the
analysis of Fortran binaries is DW_TAG_string_type, at least for now.

The only place where we need to handle the DW_TAG_string_type is in
the DIE canonicalizer.  Basically, with this patch we now consider all
the DW_TAG_string_type DIEs as being equal.  This seems good enough for
ABIs purpose, unless proven otherwise.

Note that this fixes PR23492 and PR23905.

	* src/abg-dwarf-reader.cc (die_pretty_print_type): Support
	DW_TAG_string_type DIEs.  They all have the same representation
	for now.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>